API/requirements.txt
astroid==2.3.3
certifi==2019.9.11
Click==7.0
colorama==0.4.1
dnspython==1.16.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.4.4
pymongo==3.9.0
python-dotenv==0.10.3
six==1.13.0
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
